requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
amqp==5.1.1
aniso8601==9.0.1
async-timeout==4.0.2
attrs==22.2.0
base58==2.1.1
bcrypt==4.0.1
billiard==3.6.4.0
bitarray==2.7.2
blinker==1.5
celery==5.2.7
certifi==2022.12.7
cffi==1.16.0
charset-normalizer==3.0.1
ckzg==1.0.2
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==38.0.3
cytoolz==0.12.1
dnspython==2.3.0
eth-account==0.11.2
eth-hash==0.5.1
eth-keyfile==0.8.1
eth-keys==0.5.1
eth-rlp==1.0.1
eth-typing==4.3.0
eth-utils==4.1.1
eth_abi==5.1.0
Flask==2.0.2
Flask-HTTPAuth==4.7.0
Flask-PyMongo==2.3.0
flask-redis==0.4.0
Flask-RESTful==0.3.9
Flask-Script==2.0.5
flower==1.2.0
frozenlist==1.3.3
gunicorn==20.1.0
hexbytes==0.3.0
httplib2==0.21.0
humanize==4.6.0
idna==3.4
importlib-metadata==6.0.0
ipfshttpclient==0.8.0a2
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
kombu==5.2.4
lru-dict==1.2.0
MarkupSafe==2.1.2
marshmallow==3.16.0
motor==3.1.1
msgpack==1.0.4
multiaddr==0.0.9
multidict==6.0.4
netaddr==0.8.0
packaging==23.0
parsimonious==0.10.0
prometheus-client==0.16.0
prompt-toolkit==3.0.36
protobuf==5.27.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.17
pydash==5.0.0
PyJWT==2.6.0
pymongo==4.3.3
pyparsing==3.0.9
pyrsistent==0.19.3
python-dotenv==0.16.0
python-slugify==8.0.1
pytz==2022.7.1
pyunormalize==15.1.0
redis==4.5.5
redlock-py==1.0.8
regex==2024.5.15
requests==2.28.2
rlp==4.0.1
rsa==4.9
sentry-sdk==0.10.2
six==1.16.0
text-unidecode==1.3
toolz==0.12.0
tornado==6.2
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==1.26.14
varint==1.0.2
vine==5.0.0
wcwidth==0.2.6
web3==6.19.0
websockets==12.0
Werkzeug==2.2.3
yarl==1.8.2
zipp==3.13.0
zope.event==4.6
zope.interface==5.5.2
